NEWRELIC-5287 moved out the registration of the destroy hook to reduce cognitive complexity
