+++ lib/instrumentation/core/async-hooks.js
@@ -5,8 +5,6 @@
 
 'use strict'
 
-/* eslint sonarjs/cognitive-complexity: ["error", 18] -- TODO: https://issues.newrelic.com/browse/NEWRELIC-5252 */
-
 const logger = require('../../logger').child({ component: 'async_hooks' })
 const asyncHooks = require('async_hooks')
 
@@ -26,7 +24,8 @@ function initialize(agent, shim) {
   const segmentMap = new Map()
   module.exports.segmentMap = segmentMap
 
-  const hookHandlers = getPromiseResolveStyleHooks(segmentMap, agent, shim)
+  const hookHandlers = getHookHandlers(segmentMap, agent, shim)
+  maybeRegisterDestroyHook(segmentMap, agent, hookHandlers)
 
   const hook = asyncHooks.createHook(hookHandlers)
   hook.enable()
@@ -38,8 +37,17 @@ function initialize(agent, shim) {
   return true
 }
 
-function getPromiseResolveStyleHooks(segmentMap, agent, shim) {
-  const hooks = {
+/**
+ * Registers the async hooks events
+ *
+ * Note: The init only fires when the type is PROMISE.
+ *
+ * @param {Map} segmentMap map of async ids and segments
+ * @param {Agent} agent New Relic APM agent
+ * @param {Shim} shim instance of shim
+ */
+function getHookHandlers(segmentMap, agent, shim) {
+  return {
     init: function initHook(id, type, triggerId) {
       if (type !== 'PROMISE') {
         return
@@ -101,17 +109,23 @@ function getPromiseResolveStyleHooks(segmentMap, agent, shim) {
       }
     }
   }
+}
 
-  // Clean up any unresolved promises that have been destroyed.
-  // This defaults to true but does have a significant performance impact
-  // when customers have a lot of promises.
-  // See: https://github.com/newrelic/node-newrelic/issues/760
+/**
+ * Adds the destroy async hook event that will lean up any unresolved promises that have been destroyed.
+ * This defaults to true but does have a significant performance impact
+ * when customers have a lot of promises.
+ * See: https://github.com/newrelic/node-newrelic/issues/760
+ *
+ * @param {Map} segmentMap map of async ids and segments
+ * @param {Agent} agent New Relic APM agent
+ * @param {object} hooks async-hook events
+ */
+function maybeRegisterDestroyHook(segmentMap, agent, hooks) {
   if (agent.config.feature_flag.unresolved_promise_cleanup) {
     logger.info('Adding destroy hook to clean up unresolved promises.')
     hooks.destroy = function destroyHandler(id) {
       segmentMap.delete(id)
     }
   }
-
-  return hooks
 }
