+++ lib/instrumentation/core/globals.js
@@ -5,7 +5,7 @@
 
 'use strict'
 
-const asyncHooks = require('./async_hooks')
+const asyncHooks = require('./async-hooks')
 const symbols = require('../../symbols')
 
 module.exports = initialize
